+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Show font icon for more font types
+  * Use 400 error on URI decode failure instead of 500
   * deps: accepts@~1.3.5
     - deps: mime-types@~2.1.18
   * deps: http-errors@~1.7.2